functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-storage
Pillow

